+++ OOP-project/src/main/java/com/group23/app/View/Sprite.java
@@ -28,13 +28,13 @@ public class Sprite extends JLabel{
             } else if (color.equals(Color.decode("#fc1723"))) {
                 setIcon(GameWindow.loadScaledImage(laserBasePath + "#fc1723.png", modelObject.getWidth(), modelObject.getHeight()));
             }
-            setBounds(modelObject.getX(), modelObject.getY(), modelObject.getWidth(), modelObject.getHeight());
+            setBounds((int)modelObject.getX(), (int)modelObject.getY(), modelObject.getWidth(), modelObject.getHeight());
         }
 
         // Logic for painting the Player
         else if (modelObject instanceof Player) {
             setIcon(GameWindow.loadScaledImage("OOP-project/src/main/java/com/group23/app/View/Images/Images/9Bresize.png", modelObject.getWidth(), modelObject.getHeight()));
-            setBounds(modelObject.getX(), modelObject.getY(), modelObject.getWidth(), modelObject.getHeight());
+            setBounds((int)modelObject.getX(), (int)modelObject.getY(), modelObject.getWidth(), modelObject.getHeight());
         }
 
         setBackground(Color.white);
